feat: 기존 노선에 거리 추가

+++ src/main/java/nextstep/subway/dto/LineRequest.java
@@ -7,7 +7,7 @@ public class LineRequest {
     private String color;
     private Long upStationId;
     private Long downStationId;
-    private Long distance;
+    private Integer distance;
 
     public String getName() {
         return name;
@@ -25,7 +25,7 @@ public class LineRequest {
         return downStationId;
     }
 
-    public Long getDistance() {
+    public Integer getDistance() {
         return distance;
     }
